Move sale_conventer to tests directory as is used only in tests
