+++ saleor/graphql/discount/tests/mutations/test_sale_bulk_delete.py
@@ -5,7 +5,7 @@ import pytest
 
 from .....discount.error_codes import DiscountErrorCode
 from .....discount.models import Promotion, Sale, SaleChannelListing
-from .....discount.sale_converter import convert_sales_to_promotions
+from .....discount.tests.sale_converter import convert_sales_to_promotions
 from ....tests.utils import get_graphql_content
