+++ saleor/graphql/discount/tests/queries/test_sales_pagination.py
@@ -6,7 +6,7 @@ from freezegun import freeze_time
 
 from .....discount import DiscountValueType
 from .....discount.models import Promotion, Sale, SaleChannelListing
-from .....discount.sale_converter import convert_sales_to_promotions
+from .....discount.tests.sale_converter import convert_sales_to_promotions
 from ....tests.utils import get_graphql_content
