+++ saleor/tests/fixtures.py
@@ -5635,7 +5635,7 @@ def rule_info(
 
 @pytest.fixture
 def promotion_converted_from_sale(sale):
-    from ..discount.sale_converter import convert_sales_to_promotions
+    from ..discount.tests.sale_converter import convert_sales_to_promotions
 
     convert_sales_to_promotions()
     return Promotion.objects.filter(old_sale_id=sale.id).last()
@@ -5643,7 +5643,7 @@ def promotion_converted_from_sale(sale):
 
 @pytest.fixture
 def promotion_converted_from_sale_with_empty_predicate():
-    from ..discount.sale_converter import convert_sales_to_promotions
+    from ..discount.tests.sale_converter import convert_sales_to_promotions
 
     sale = Sale.objects.create(name="Sale with no rules", type=DiscountValueType.FIXED)
     convert_sales_to_promotions()
